pyproject.toml
[tool.poetry]
name = "repl_python3_jovian-careers-website"
version = "0.1.0"
description = ""
authors = ["aakashns <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.1.2"
sqlalchemy = "^2.0.17"
pymysql = "^1.1.0"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
